rename api to dto
