+++ src/main/java/main/dto/request/SettingsRequest.java
@@ -1,4 +1,4 @@
-package main.api.request;
+package main.dto.request;
 
 import lombok.Getter;
 import lombok.Setter;
