+++ src/main/java/main/service/PostService.java
@@ -1,7 +1,7 @@
 package main.service;
 
-import main.api.responses.PostResponseForList;
-import main.api.responses.PostsResponse;
+import main.dto.responses.PostResponseForList;
+import main.dto.responses.PostsResponse;
 import main.model.Posts;
 import main.repositories.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
